Update README.md
Add note about how this only works with CODAP interactives with scaling enabled

It would be nice if this worked with table interactives too and with CODAP interactives without scaling. Technically it should at least work for the non scaling CODAP interactives. My guess is that the scaling slows things down enough that some race condition is avoided.