Create comment list DB structure and view
When a user makes a comment (posting not available yet in this commit)
it is stored under the user's own userid in the DB. This way the simple
Firestore rules will give a user ownership over their comments.

This requires a little sorting of comments into the sender's bins once
we receive them, but this isn't so bad.

Since we want to keep comments even after a user unshares, unsharing no
longer deletes the user's document but instead just blanks the iframeUrl.