Use a ref instead of state inside mouseMove
Because of the way `onMouseMove` is defined now, the reference to
`topHeight` gets locked in a closure, which means that the `topHeight`
referenced by `onMouseDown` and the one referenced by `onMouseMove` are
different. `onMouseMove` never sees its own updates via `setTopHeight`,
and every time it is called it reads the same value.

This modifies `onMouseMove` so it references a `ref`, which does get
updated, but it still sets state so that we correctly trigger a render.