fix: Sharing plugin modal z-index issues [PT-185637498]
After trying a few different approaches to fixing the sharing modals being hidden in the grid layout in AP due to the grid element's stacking context this fix switches to rendering the modals via createPortal in a dynamically inserted div at the end of the body for each render.

This approach ensures that the modals "escape" the grid layout.